refactor(general): Refactor helper package and respective imports
